The card label logger script defined
Card logger script has been defined and is picking up the number-label of the card it is shown. TODO: Get rid of the code that prevents logging of repeats as we WANT repeats in this scenario. Desired output of the logger would be an array or csv containing the first 100 or 1000 the logged label numbers so we can perform data analysis on them.